+++ js/glotdict.js
@@ -89,4 +89,6 @@ gd_wait_table_alter();
 
 gd_remove_layover();
 
-jQuery('.gp-content').css('max-width', '85%');
+if (jQuery('.gp-content #bulk-actions-toolbar-top').length > 0) {
+  jQuery('.gp-content').css('max-width', '85%');
+}